+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - added: (ADA) `isStakeTx` to `EdgeTransaction.otherParams` for staking transaction signing
+- added: (ADA) staking related methods to `EdgeCurrencyEngine.otherMethods`
 - fixed: (Solana) Properly validate custom token addresses
 
 ## 4.22.0 (2024-08-30)